See full commit message for list of changes:
Confirm and Alert modals created programmatically now accept a backdropClickable prop
Confirm and Alert modals no longer need ‘type’ passed as prop. Default will be ‘info’ if not passed.
The only required props for Confirm and Alert modals now are ‘title’ and ‘message’. As it should be.
VaRange now correctly updates itself when the value it is bound to changes.
VaRange now correctly updates itself when the page is resized and, thus, the range slider width could possibly change.
Modal was emitting hide and show twice. Whoops. This is now fixed.