Please see full commit message for list of changes.
Many, many changes. None of which should break any existing projects.

Removed VaBars - it was not being used anyways.

Due to the complexity of the LayoutManager, a new component has been introduced to help with its construction: VaLayoutManagerConfig. Please see the documentation page (https://vue-atlas.com/documentation/layoutmanagerconfig) for a guide on how it is meant to be used.

VaLayoutManager has quite a few new props (use VaLayoutManagerConfig to see examples of all of this):
- desktopMargin
- desktopMinimumWidth
- topbarPriority
- sidebarPriority
- minibarPriority
- split
- reverse
- bgColor
- pageBgColor

Other, smaller, misc. changes:

_timepicker.scss - slightly increased border radius for timepicker dropdown.
_modal.scss - overflow: hidden changed to overflow: scroll. Modals extending beyond page bounds can be scrolled now. Increased margin-bottom.
_layout.scss - changed breakpoints at which the va-columns will resize.
VaModal: new backdrop prop allows you to specify if backdrop is shown or not. Modals can also now be stacked on top of each other. Instantiate as many modals as you’d like to see this effect.